+++ app/routes/tasks.py
@@ -5,6 +5,7 @@ from app.models.project import Project
 from app.schemas.task_schema import TaskSchema
 from flask_jwt_extended import get_jwt_identity, get_jwt, jwt_required
 from app.utils.security import role_required
+from app.utils.custom_exceptions import Forbidden, ResourceNotFound, BadRequest, UnprocessableEntity
 
 # Input: task_name, description, project_id
 
@@ -19,17 +20,24 @@ tasks_schema = TaskSchema(many = True)
 @jwt_required()
 @role_required(["user", "admin"])
 def get_tasks(project_id: int):
-    project = Project.query.get_or_404(project_id)      # Get Project if exists
+    project = Project.query.filter_by(id = project_id).first()      # Get Project if exists
+    if not project:
+        raise ResourceNotFound("Project not found.")
+      
     current_user_id = int(get_jwt_identity())
     claims = get_jwt()
         
     if claims["role"] != "admin" and project.owner_id != current_user_id:    # Make sure proper ownership
-        return jsonify(message = "Access Denied"), 403
+        raise Forbidden("Access Denied")  
+    
+    tasks = Task.query.filter_by(project_id = project_id).all()
     
     if claims["role"] == "admin": 
         tasks = Task.query.all()
-
-    tasks = Task.query.filter_by(project_id = project_id).all()
+        
+    if not tasks:
+        return jsonify(message = "No tasks for this project this time."), 404
+    
     return tasks_schema.jsonify(tasks), 200
 
 
@@ -38,17 +46,22 @@ def get_tasks(project_id: int):
 @jwt_required()
 @role_required(["user", "admin"])
 def get_task(project_id: int, task_id: int):
-    project = Project.query.get_or_404(project_id)      # Get Project if exists
+    project = Project.query.filter_by(id = project_id).first()      # Get Project if exists
+    if not project:
+        raise ResourceNotFound("Project not found.")
+    
     current_user_id = int(get_jwt_identity())
     claims = get_jwt()
     
     if claims["role"] != "admin" and project.owner_id != current_user_id:
-        return jsonify(message = "Access denied"), 403
+        raise Forbidden("Access Denied")
     
-    task = Task.query.get_or_404(task_id) # Check if specific tasks exists
+    task = Task.query.filter_by(id = task_id).first() # Check if specific tasks exists
+    if not task:
+        raise ResourceNotFound("Task not found.")
     
     if task.project_id != project_id:
-        return jsonify(message = "Access denied"), 403
+        raise Forbidden("Access Denied")
     
     return task_schema.jsonify(task), 200
     
@@ -58,19 +71,22 @@ def get_task(project_id: int, task_id: int):
 @jwt_required()
 @role_required(["user", "admin"])
 def create_task(project_id: int):
-    project = Project.query.get_or_404(project_id)      # Get Project if exists
+    project = Project.query.filter_by(id = project_id).first()      # Get Project if exists
+    if not project:
+        raise ResourceNotFound("Project not found.")
+    
     current_user_id = int(get_jwt_identity())
     claims = get_jwt()
     json_data = request.get_json()
     
     # Check for input
     if not json_data:
-        return jsonify(message = "Invalid data"), 400
+        raise BadRequest("No data provided.")
     
     # Validate Input
     errors = task_schema.validate(json_data)
     if errors:
-        return jsonify(errors), 400
+        raise UnprocessableEntity(errors)
     
     # Check ownership
     if claims["role"] == "admin" or project.owner_id == current_user_id:
@@ -86,7 +102,7 @@ def create_task(project_id: int):
         db.session.commit()
         return task_schema.jsonify(task), 201
     
-    return jsonify(message = "Access Denied"), 403
+    raise Forbidden("Access Denied")
 
 
 # PUT /projects/<project_id>/tasks/<task_id>
@@ -94,24 +110,29 @@ def create_task(project_id: int):
 @jwt_required()
 @role_required(["user", "admin"])
 def update_task(project_id: int, task_id: int):
-    project = Project.query.get_or_404(project_id)      # Get Project if exists
+    project = Project.query.filter_by(id = project_id).first()      # Get Project if exists
+    if not project:
+        raise ResourceNotFound("Project not found.")
+    
     current_user_id = int(get_jwt_identity())
     claims = get_jwt()
     json_data = request.get_json()
     
     if claims["role"] != "admin" and project.owner_id != current_user_id:
-        return jsonify(message = "Access Denied"), 403
+        raise Forbidden("Access Denied")
     
     if not json_data:
-        return jsonify(message = "No data provided."), 400   
+        raise BadRequest("No data provided.")
     
     # Validate input
     errors = task_schema.validate(json_data)
     if errors:
-        return jsonify(errors), 400
+        raise UnprocessableEntity(errors)
     
     # Check if task exists
-    task = Task.query.get_or_404(task_id) 
+    task = Task.query.filter_by(id = task_id).first()
+    if not task:    
+        raise ResourceNotFound("Task not found.") 
     
     # Deserialize data
     data = task_schema.load(json_data)
@@ -127,17 +148,22 @@ def update_task(project_id: int, task_id: int):
 @jwt_required()
 @role_required(["user", "admin"])
 def remove_task(project_id: int, task_id: int):
-    project = Project.query.get_or_404(project_id)      # Get Project if exists
+    project = Project.query.filter_by(id = project_id).first()      # Get Project if exists
+    if not project:
+        raise ResourceNotFound("Project not found.")
+    
     current_user_id = int(get_jwt_identity())
     claims = get_jwt()
     
     if claims["role"] != "admin" and project.owner_id != current_user_id:
-        return jsonify(message = "Access Denied"), 403
+        raise Forbidden("Access Denied")
     
-    task = Task.query.get_or_404(task_id)
+    task = Task.query.filter_by(id = task_id).first()
+    if not task:    
+        raise ResourceNotFound("Task not found.") 
     
     if task.project_id != project_id:
-        return jsonify(message = "Access denied"), 403
+        raise Forbidden("Access Denied")
     
     db.session.delete(task)
     db.session.commit()
